backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
motor==3.6.0
email-validator==2.1.1
